extract module series
